Create importDetails model
